ai-engine/requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.7.2
pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
